+++ Taskfile.py
@@ -1,4 +1,3 @@
-import os
 import zipimport
 from functools import partial
 from pathlib import Path
@@ -34,16 +33,25 @@ yaml = importer.load_module("yaml")
 
 class RunTaskCommand(sublime_plugin.WindowCommand):
     def run(self):
-        taskfile = open(cwd / "Taskfile.yml")
-        res = yaml.load(taskfile, Loader=yaml.FullLoader)
+        folders = self.window.folders()
+        if not folders or len(folders) > 1:
+            self.window.status_message("Only one folder per project is supported.")
+            return
+        working_dir_path = Path(folders[0])
+        taskfile_path = working_dir_path / "Taskfile.yml"
+        res = yaml.load(taskfile_path.open(), Loader=yaml.FullLoader)
         items = get_quick_panel_items(res)
-        on_done = partial(run_task_by_index, self.window, items)
+        on_done = partial(run_task_by_index, self.window, items, str(working_dir_path))
         self.window.show_quick_panel(items, on_done, sublime.MONOSPACE_FONT)
 
 
-def run_task_by_index(window, quick_panel_items, index):
+def run_task_by_index(window, quick_panel_items, working_dir, index):
+    if index < 0:
+        return
     task_name = quick_panel_items[index].trigger
-    window.run_command("exec", args={"shell_cmd": f"task {task_name}"})
+    window.run_command(
+        "exec", args={"shell_cmd": f"task {task_name}", "working_dir": working_dir}
+    )
 
 
 def get_quick_panel_items(taskfile):